requirements.txt
asgiref==3.3.2
boto3==1.11.9
botocore==1.14.9
certifi==2019.11.28
chardet==3.0.4
debugpy==1.5.1
Django==3.2.8
django-json-widget==1.1.1
django-cleanup==4.0.0
django-extensions==2.2.6
django-pipeline==2.0.7
django-storages==1.9
djangorestframework==3.11.0
docutils==0.15.2
idna==2.8
jmespath==0.9.4
mysqlclient==1.4.6
pdf2image==1.14.0
-e git+https://github.com/gofr/Pillow.git@91ac066fbce99ab2166cfb0e2af4ae8513d3e959#egg=Pillow
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
s3transfer==0.3.2
six==1.14.0
sqlparse==0.3.0
urllib3==1.25.8
django-simple-menu==1.2.1
gunicorn==20.1.0
pymysql==1.0.2
python-poppler==0.2.2
django-js-urls==0.1.1
colorlog==4.8.0
django-cors-headers==3.11.0
-e git+https://${BITBUCKET_USER}:${BITBUCKET_APP_PASSWORD}@bitbucket.org/badilsrl/connector-library/src/v1.8.7/#egg=connector-library
pre-commit==2.15.0
pylint==2.11.1
black==21.9b0
django-configurations==2.3.1
